changed restart workflow

+++ mesheryctl/internal/cli/root/system/restart.go
@@ -67,41 +67,41 @@ func restart() error {
 		}
 
 	case "kubernetes":
-		// create an kubernetes client
-		client, err := meshkitkube.New([]byte(""))
-
+		running, err := utils.IsMesheryRunning(currPlatform)
 		if err != nil {
 			return err
 		}
-
-		// Create a pod interface for the MesheryNamespace
-		podInterface := client.KubeClient.CoreV1().Pods(utils.MesheryNamespace)
-
-		// List the pods in the MesheryNamespace
-		podList, err := podInterface.List(context.TODO(), v1.ListOptions{})
-		if err != nil {
-			return err
-		}
-
-		// List all the pods similar to kubectl get pods -n MesheryNamespace
-		for _, pod := range podList.Items {
-			// Get the values from the pod status
-			name := pod.GetName()
-			log.Info("Deleting pod ", name)
-			err := client.KubeClient.CoreV1().Pods(utils.MesheryNamespace).Delete(context.TODO(), name, v1.DeleteOptions{})
-			if err != nil {
-				log.Fatal(err)
-			}
-			log.Info("Restarting pod ", name)
-		}
-
-		ok, err := utils.IsMesheryRunning(currPlatform)
-		if err != nil {
-			return err
-		}
-		if !ok {
+		if !running { // Meshery is not running
 			if err := start(); err != nil {
 				return errors.Wrap(err, utils.SystemError("Failed to restart Meshery"))
+			}
+		} else {
+			// create a kubernetes client
+			client, err := meshkitkube.New([]byte(""))
+
+			if err != nil {
+				return err
+			}
+
+			// Create a pod interface for the MesheryNamespace
+			podInterface := client.KubeClient.CoreV1().Pods(utils.MesheryNamespace)
+
+			// List the pods in the MesheryNamespace
+			podList, err := podInterface.List(context.TODO(), v1.ListOptions{})
+			if err != nil {
+				return err
+			}
+
+			// List all the pods similar to kubectl get pods -n MesheryNamespace
+			for _, pod := range podList.Items {
+				// Get the values from the pod status
+				name := pod.GetName()
+				log.Info("Deleting pod ", name)
+				err := client.KubeClient.CoreV1().Pods(utils.MesheryNamespace).Delete(context.TODO(), name, v1.DeleteOptions{})
+				if err != nil {
+					log.Fatal(err)
+				}
+				log.Info("Restarting pod ", name)
 			}
 		}
 	}
